moved migration to front, fixing up data dependant tests (10/20 fail)
