+++ cypress/integration/admin/addingATrust.js
@@ -1,9 +1,13 @@
+import { whenIClickLogOut } from "../commonSteps";
+import {
+  GivenIAmLoggedInAsAnAdmin,
+  ThenISeeTheSiteAdministrationPage,
+  WhenIClickToReturnToSiteAdministration,
+} from "./adminCommonSteps";
+
 describe("As an admin, I want to add a trust so that a trust can use the virtual visits service.", () => {
-  before(() => {
-    // reset and seed the database
-    cy.exec(
-      "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-    );
+  after(() => {
+    whenIClickLogOut();
   });
 
   it("allows an admin to add a trust", () => {
@@ -33,14 +37,6 @@ describe("As an admin, I want to add a trust so that a trust can use the virtual
     ThenISeeErrors();
   });
 
-  // Allows an admin to add a trust
-  function GivenIAmLoggedInAsAnAdmin() {
-    cy.visit(Cypress.env("baseUrl") + "/admin/login");
-    cy.get("input[name=code]").type(Cypress.env("validAdminCode"));
-    cy.get("input[name=password]").type(Cypress.env("validAdminPassword"));
-    cy.get("button").contains("Log in").click();
-  }
-
   function WhenIClickOnAddATrust() {
     cy.get("a.nhsuk-action-link__link").contains("Add a trust").click();
   }
@@ -63,14 +59,6 @@ describe("As an admin, I want to add a trust so that a trust can use the virtual
 
   function ThenISeeTheTrustIsAdded() {
     cy.get("h1").should("contain", "Bow Trust has been added");
-  }
-
-  function WhenIClickToReturnToSiteAdministration() {
-    cy.get("a").contains("Return to site administration").click();
-  }
-
-  function ThenISeeTheSiteAdministrationPage() {
-    cy.get("h1").should("contain", "Site administration");
   }
 
   function AndISeeTheAddedTrust() {
