+++ cypress/integration/admin/editingATrust.js
@@ -1,9 +1,12 @@
+import { whenIClickLogOut } from "../commonSteps";
+import {
+  GivenIAmLoggedInAsAnAdmin,
+  ThenISeeTheSiteAdministrationPage,
+} from "./adminCommonSteps";
+
 describe("As an admin, I want to edit a trust so that I can update the details of a trust.", () => {
-  before(() => {
-    // reset and seed the database
-    cy.exec(
-      "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-    );
+  after(() => {
+    whenIClickLogOut();
   });
 
   it("allows an admin to edit a trust", () => {
@@ -23,14 +26,6 @@ describe("As an admin, I want to edit a trust so that I can update the details o
     ThenISeeTheSiteAdministrationPage();
     AndISeeTheUpdatedTrust();
   });
-
-  // Allows an admin to edit a trust
-  function GivenIAmLoggedInAsAnAdmin() {
-    cy.visit(Cypress.env("baseUrl") + "/admin/login");
-    cy.get("input[name=code]").type(Cypress.env("validAdminCode"));
-    cy.get("input[name=password]").type(Cypress.env("validAdminPassword"));
-    cy.get("button").contains("Log in").click();
-  }
 
   function WhenIClickToEditATrust() {
     cy.get("a.nhsuk-link").contains("Edit Test Trust").click();
@@ -54,10 +49,6 @@ describe("As an admin, I want to edit a trust so that I can update the details o
 
   function WhenIClickToReturnToSiteAdministration() {
     cy.get("a").contains("Return to site administration").click();
-  }
-
-  function ThenISeeTheSiteAdministrationPage() {
-    cy.get("h1").should("contain", "Site administration");
   }
 
   function AndISeeTheUpdatedTrust() {
