+++ cypress/integration/admin/loggingInAndOut.js
@@ -1,9 +1,8 @@
+import { whenIClickLogOut } from "../commonSteps";
+
 describe("As an admin, I want to log in so that I can access the service.", () => {
-  before(() => {
-    // reset and seed the database
-    cy.exec(
-      "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-    );
+  after(() => {
+    // this one does not need to logout as all tests end in already logged out state
   });
 
   it("allows an admin to log in and out", () => {
@@ -18,7 +17,7 @@ describe("As an admin, I want to log in so that I can access the service.", () =
 
     cy.audit();
 
-    WhenIClickLogOut();
+    whenIClickLogOut();
     ThenISeeTheAdminLogInPage();
   });
 
@@ -55,10 +54,6 @@ describe("As an admin, I want to log in so that I can access the service.", () =
   function ThenISeeTheAdminHomePage() {
     cy.contains("There is a problem").should("not.be.visible");
     cy.contains("Site administration").should("be.visible");
-  }
-
-  function WhenIClickLogOut() {
-    cy.get("a.nhsuk-header__navigation-link").contains("Log out").click();
   }
 
   function ThenISeeTheAdminLogInPage() {
