+++ cypress/integration/commonSteps.js
@@ -1,0 +1,5 @@
+module.exports = {
+  whenIClickLogOut: () => {
+    cy.get("a.nhsuk-header__navigation-link").contains("Log out").click();
+  },
+};
