+++ cypress/integration/trust-admin/addingAWard.js
@@ -1,9 +1,15 @@
+import { whenIClickLogOut } from "../commonSteps";
+import {
+  GivenIAmLoggedInAsATrustAdmin,
+  ThenISeeTheHospitalsPage,
+  ThenISeeThePageForTheHospital,
+  WhenIClickHospitalsOnTheNavigationBar,
+  WhenIClickToReturnToThePageForTheHospital,
+} from "./trustAdminCommonSteps";
+
 describe("As a trust admin, I want to add a ward so that ward staff can book virtual visits.", () => {
-  before(() => {
-    // reset and seed the database
-    cy.exec(
-      "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-    );
+  after(() => {
+    whenIClickLogOut();
   });
 
   it("allows a trust admin to add a ward", () => {
@@ -43,30 +49,8 @@ describe("As a trust admin, I want to add a ward so that ward staff can book vir
     ThenISeeErrors();
   });
 
-  // Allows a trust admin to add a hospital
-  function GivenIAmLoggedInAsATrustAdmin() {
-    cy.visit(Cypress.env("baseUrl") + "/trust-admin/login");
-
-    cy.get("input[name=code]").type(Cypress.env("validTrustAdminCode"));
-    cy.get("input[name=password]").type(Cypress.env("validTrustAdminPassword"));
-
-    cy.get("button").contains("Log in").click();
-  }
-
-  function WhenIClickHospitalsOnTheNavigationBar() {
-    cy.get("a.nhsuk-header__navigation-link").contains("Hospitals").click();
-  }
-
-  function ThenISeeTheHospitalsPage() {
-    cy.get("h1").should("contain", "Hospitals");
-  }
-
   function WhenIClickOnAHospital() {
     cy.get("a").contains("View Test Hospital").click();
-  }
-
-  function ThenISeeThePageForTheHospital() {
-    cy.get("h1").should("contain", "Test Hospital");
   }
 
   function WhenIClickOnAddAWard() {
@@ -89,10 +73,6 @@ describe("As a trust admin, I want to add a ward so that ward staff can book vir
 
   function ThenISeeTheWardIsAdded() {
     cy.get("h1").should("contain", "Glimmer Ward has been added");
-  }
-
-  function WhenIClickToReturnToThePageForTheHospital() {
-    cy.get("a").contains("Return to Test Hospital").click();
   }
 
   function AndISeeTheAddedWard() {
