+++ cypress/integration/trust-admin/deletingAWard.js
@@ -1,9 +1,15 @@
+import {
+  GivenIAmLoggedInAsATrustAdmin,
+  ThenISeeTheHospitalsPage,
+  ThenISeeThePageForTheHospital,
+  WhenIClickHospitalsOnTheNavigationBar,
+  WhenIClickToReturnToThePageForTheHospital,
+} from "./trustAdminCommonSteps";
+import { whenIClickLogOut } from "../commonSteps";
+
 describe("As a trust admin, I want to delete a ward so that it can no longer use the virtual visits service.", () => {
-  before(() => {
-    // reset and seed the database
-    cy.exec(
-      "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-    );
+  after(() => {
+    whenIClickLogOut();
   });
 
   it("allows a trust admin to delete a ward", () => {
@@ -26,30 +32,8 @@ describe("As a trust admin, I want to delete a ward so that it can no longer use
     AndIDoNotSeeTheDeletedWard();
   });
 
-  // Allows a trust admin to delete a ward
-  function GivenIAmLoggedInAsATrustAdmin() {
-    cy.visit(Cypress.env("baseUrl") + "/trust-admin/login");
-
-    cy.get("input[name=code]").type(Cypress.env("validTrustAdminCode"));
-    cy.get("input[name=password]").type(Cypress.env("validTrustAdminPassword"));
-
-    cy.get("button").contains("Log in").click();
-  }
-
-  function WhenIClickHospitalsOnTheNavigationBar() {
-    cy.get("a.nhsuk-header__navigation-link").contains("Hospitals").click();
-  }
-
-  function ThenISeeTheHospitalsPage() {
-    cy.get("h1").should("contain", "Hospitals");
-  }
-
   function WhenIClickOnAHospital() {
     cy.get("a").contains("View Test Hospital").click();
-  }
-
-  function ThenISeeThePageForTheHospital() {
-    cy.get("h1").should("contain", "Test Hospital");
   }
 
   function WhenIClickToDeleteAWard() {
@@ -69,10 +53,6 @@ describe("As a trust admin, I want to delete a ward so that it can no longer use
 
   function ThenISeeTheWardIsDeleted() {
     cy.get("h1").should("contain", "Test Ward One has been deleted");
-  }
-
-  function WhenIClickToReturnToThePageForTheHospital() {
-    cy.get("a").contains("Return to Test Hospital").click();
   }
 
   function AndIDoNotSeeTheDeletedWard() {
