+++ cypress/integration/wards/cancellingAVisit.js
@@ -1,9 +1,13 @@
+import {
+  GivenIAmLoggedInAsAWardStaff,
+  WhenIClickOnAVirtualVisit,
+  WhenIClickReturnToVirtualVisits,
+} from "./wardCommonSteps";
+import { whenIClickLogOut } from "../commonSteps";
+
 describe("As a ward staff, I want to cancel a virtual visit so that the visit cannot be started.", () => {
-  before(() => {
-    // reset and seed the database
-    cy.exec(
-      "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-    );
+  after(() => {
+    whenIClickLogOut();
   });
 
   it("allows a ward staff to cancel a virtual visit", () => {
@@ -20,16 +24,6 @@ describe("As a ward staff, I want to cancel a virtual visit so that the visit ca
     WhenIClickReturnToVirtualVisits();
     ThenIDoNotSeeTheVirtualVisit();
   });
-
-  function GivenIAmLoggedInAsAWardStaff() {
-    cy.visit(Cypress.env("baseUrl"));
-    cy.get("input").type(Cypress.env("validWard"));
-    cy.get("button").contains("Log in").click();
-  }
-
-  function WhenIClickOnAVirtualVisit() {
-    cy.get("summary.nhsuk-details__summary").contains("Alice").click();
-  }
 
   function AndIClickOnCancel() {
     cy.get("summary.nhsuk-details__summary")
@@ -58,10 +52,6 @@ describe("As a ward staff, I want to cancel a virtual visit so that the visit ca
 
   function ThenISeeTheVirtualVisitIsCancelled() {
     cy.get("h1").should("contain", "Virtual visit cancelled");
-  }
-
-  function WhenIClickReturnToVirtualVisits() {
-    cy.get("a").contains("Return to virtual visits").click();
   }
 
   function ThenIDoNotSeeTheVirtualVisit() {
