--- cypress/support/databaseSetup.js
@@ -1,3 +1,0 @@
-cy.exec(
-  "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed"
-);
