+++ cypress/support/index.js
@@ -16,6 +16,5 @@
 // Import commands.js using ES2015 syntax:
 import "./commands";
 import "./audit";
-// import "./databaseSetup";
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
