+++ package.json
@@ -18,6 +18,8 @@
     "test:server": "APP_ENV=test NODE_ENV=development PORT=3001 node server.js",
     "test:coverage": "jest --silent --coverage",
     "test:ci": "npm run format:check && npm run test:contract -- --runInBand && npm run test -- --runInBand",
+    "test:migrate": "npm run dbmigratetest reset && npm run dbmigratetest up && npm run db:seed",
+    "test:e2e2": "npm run test:migrate && npm run test:e2e",
     "build": "next build",
     "dev": "NODE_ENV=development node server.js",
     "start": "NODE_ENV=production node server.js",
